requirements.txt
openai==0.28
faker
pillow
flask
scikit-learn
numpy
pytesseract